refactor types and add data explorer view enum
